ifcparser/requirements.txt
bimdata-api-client>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0 